improve disk usage of zap (#15)
Breaking change, to be released as new major version (v13.0.0)

One of the optimizations in v12 was to encode a special marker
value when a field/term had no location information stored.
Unfortunately that marker was math.MaxUint64, which when
encoded via Uvarint requires 10 bytes.  In practice, the number
of bytes required went up, not down as a result of this change.

Now, we use 0 as that special marker value.  This has been
determined to be safe as 0 could never be the actual address
of term location information, as there are other mandatory
sections earlier in the file (stored field index).