New norm disk format for zap
The current norm encoding consistently makes a footprint of
size greater 5bytes on disk irrespective of the actual field length.

This size bloating could be saved if we directly encode the field
length directly to disk and decode it correctly during the read time.

More than 30% reduction in index size is observed with this change

The current change also restricts the changes to zap itself.

As this fix changes the meaning of data present on disk,
it mandates a zap version bump.

There could be variations in norm's precison stemming from the
floating point precision loss between uint64 to float64 conversions
on some higher field length values, but trading it off for the simple
conversion logic.

Any alternate suggestions appreciated.