Enabling configurable freq/norm with storage
With freq/norm disabled at the index layer, storage
layer no longer needs to persist both freq/norm related
information onto disk.
A special frequency value of zero/0 is interpreted as
a configuration for skipping the processing of norm
value.

So with freq/norm disabled, zap always store the freq
value of 0 along with an LSB indicating the presence
of location data.

Savings are mostly of 1 byte per occurrence of the
token as we always persist a freq value of 0 with
the location LSB.