Fixed 64-bit alignment panic in chunkedContentCoder on 32-bit platforms.
Following the fix in #147, a 'panic: unaligned 64-bit atomic operation' is
still occurring on some platforms.

As noted in #147, this commit implements the alternative fix by placing the
bytesWritten field at the top of the struct. Per the Golang sync/atomic
documentation, on ARM, 386 and 32-bit MIPS systems, this is the reliable
way to prevent these issues.

Signed-off-by: Callum Jones <cj@icj.me>