[BP] Handle zero chunk size
This is a backport of https://github.com/blevesearch/zapx/pull/265.

The existing implementation of the method
getChunkSize(...) (uint64, error), in some cases,
can return chunkSize==0 and err==nil.

The onus is on the caller to check for such possibility and handle it
properly.
Callers often use the returned chunkSize as a divisor and a zero
chunkSize lead to panic.
see https://github.com/blevesearch/zapx/issues/209

This PR intends to update the method implementation to always return an
error in case the returned chunkSize value is 0.
That way caller need to only worry about error being non-nil.

Callers which are ok with 0 chunkSize can check the returned error
against ErrChunkSizeZero