MB-69641: Add support for retrieval of multi-vector documents (#348)
- Add a new `docSearch` wrapper that allows any FAISS search function to
be executed in a loop, filtering out lower quality vectors from
duplicate documents in each iteration.
- Add `MaxDeduplicationIterations` as a global variable that can be set
as a config to place an upper bound on the maximum number of iterations
the deduplication loop will be run. The current default is `100`.
- Refactor the existing search functions to always execute inside the
wrapper, allowing the vector index to be searched multiple times until
either `k` unique documents are returned, or the index is exhausted or
we hit the `MaxMultiVectorDocSearchRetries` number of loop iterations,
instead of returning the `k` best vectors.

---------

Co-authored-by: Likith B <likith.b@couchbase.com>